+++ server/db/add_buttons_to_client.sh
@@ -15,7 +15,7 @@ elif [[ ! -n "$3" ]]; then
     echo "Example: $0 ./../.env ExistingClient ./add_buttons.csv.example"
     echo ""
     echo "The buttons CSV file"
-    echo "MUST have the header 'button_id,unit,phone_number,button_serial_number'"
+    echo "MUST have the header 'unit,phone_number,button_serial_number'"
     echo "MUST use Unix line endings (LF), or else the phone numbers will have '\r' at the end"
     echo "MUST end with a newline, or else the last button will be silently ignored"
     echo ""
@@ -41,15 +41,14 @@ else
         exit 1
     fi
 
-    while IFS=",", read -r button_id unit phone_number button_serial_number; do
+    while IFS=",", read -r unit phone_number button_serial_number; do
         if [[ "$phone_number" != "phone_number" && "$phone_number" != "" ]]; then
             echo "Adding button"
-            echo "  Button ID: $button_id"
             echo "  Display Name: $unit"
             echo "  Phone Number: $phone_number"
             echo "  Serial Number: $button_serial_number"
 
-        sudo PGPASSWORD=$PG_PASSWORD psql -U $PG_USER -h $PG_HOST -p $PG_PORT -d $PG_USER --set=sslmode=require -c "INSERT INTO buttons (button_id, button_serial_number, display_name, phone_number, client_id) VALUES ('$button_id', '$button_serial_number', '$unit', '$phone_number', '$client_id');"
+        sudo PGPASSWORD=$PG_PASSWORD psql -U $PG_USER -h $PG_HOST -p $PG_PORT -d $PG_USER --set=sslmode=require -c "INSERT INTO buttons (button_serial_number, display_name, phone_number, client_id) VALUES ('$button_serial_number', '$unit', '$phone_number', '$client_id');"
         fi
     done < $3